Tidy up Typha tests.
