+++ typha/fv-tests/server_test.go
@@ -15,16 +15,12 @@
 package fvtests_test
 
 import (
-	"bytes"
 	"context"
 	"crypto/x509"
-	"encoding/base64"
-	"encoding/binary"
 	"encoding/gob"
 	"errors"
 	"fmt"
 	"io/ioutil"
-	"math/rand"
 	"net"
 	"os"
 	"path/filepath"
@@ -35,10 +31,14 @@ import (
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
 	apiv3 "github.com/projectcalico/api/pkg/apis/projectcalico/v3"
+	"github.com/prometheus/client_golang/prometheus"
+	io_prometheus_client "github.com/prometheus/client_model/go"
+	log "github.com/sirupsen/logrus"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
 	libapiv3 "github.com/projectcalico/calico/libcalico-go/lib/apis/v3"
 	"github.com/projectcalico/calico/libcalico-go/lib/backend/api"
 	"github.com/projectcalico/calico/libcalico-go/lib/backend/encap"
-	"github.com/projectcalico/calico/libcalico-go/lib/backend/k8s/conversion"
 	"github.com/projectcalico/calico/libcalico-go/lib/backend/model"
 	calinet "github.com/projectcalico/calico/libcalico-go/lib/net"
 	. "github.com/projectcalico/calico/typha/fv-tests"
@@ -49,11 +49,6 @@ import (
 	"github.com/projectcalico/calico/typha/pkg/syncproto"
 	"github.com/projectcalico/calico/typha/pkg/syncserver"
 	"github.com/projectcalico/calico/typha/pkg/tlsutils"
-	"github.com/prometheus/client_golang/prometheus"
-	io_prometheus_client "github.com/prometheus/client_model/go"
-	log "github.com/sirupsen/logrus"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 var (
@@ -158,17 +153,6 @@ var (
 	}
 )
 
-type NoOpCallbacks struct {
-}
-
-func (n NoOpCallbacks) OnStatusUpdated(status api.SyncStatus) {
-}
-
-func (n NoOpCallbacks) OnUpdates(updates []api.Update) {
-}
-
-// Tests that rely on starting a real Typha syncserver.Server (on a real TCP port) in this process.
-// We drive the server via a real snapshot cache using the snapshot cache's function API.
 var _ = Describe("With an in-process Server", func() {
 	// We'll create this pipeline for updates to flow through:
 	//
@@ -176,7 +160,7 @@ var _ = Describe("With an in-process Server", func() {
 	//                      decoupler        filter        cache
 	//
 	var (
-		h *serverHarness
+		h *ServerHarness
 	)
 
 	BeforeEach(func() {
@@ -194,96 +178,6 @@ var _ = Describe("With an in-process Server", func() {
 		Expect(h.Server.Port()).ToNot(BeZero())
 	})
 
-	randomValue := func(n int) string {
-		v := fmt.Sprint(n, "=")
-		buf := make([]byte, 1000)
-		rand.Read(buf)
-		var outBuf bytes.Buffer
-		enc := base64.NewEncoder(base64.StdEncoding, &outBuf)
-		_, _ = enc.Write(buf)
-		v += outBuf.String()
-		return v
-	}
-	_ = randomValue
-
-	randomHex := func(length int) string {
-		buf := make([]byte, length/2)
-		rand.Read(buf)
-		return fmt.Sprintf("%x", buf)
-	}
-
-	generatePod := func(n int) *corev1.Pod {
-		namespace := fmt.Sprintf("a-namespace-name-%x", n/100)
-		var buf [8]byte
-		rand.Read(buf[:])
-		name := fmt.Sprintf("some-app-name-%d-%x", n, buf[:])
-		hostname := fmt.Sprintf("hostname%d", n/20)
-		ip := net.IP{0, 0, 0, 0}
-		binary.BigEndian.PutUint32(ip, uint32(n))
-		ip[0] = 10
-		p := &corev1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: namespace,
-				Name:      name,
-				Annotations: map[string]string{
-					"cni.projectcalico.org/containerID": randomHex(64),
-					"cni.projectcalico.org/podIP":       fmt.Sprintf("%s,/32", ip.String()),
-					"cni.projectcalico.org/podIPs":      fmt.Sprintf("%s,/32", ip.String()),
-					"junk":                              randomHex(10000),
-				},
-				Labels: map[string]string{
-					"kubernetes-topology-label": "zone-A",
-					"kubernetes-region-label":   "zone-A",
-					"owner":                     randomHex(23),
-					"oneof10":                   fmt.Sprintf("value-%d", n/10),
-					"oneof100":                  fmt.Sprintf("value-%d", n/100),
-				},
-			},
-			Spec: corev1.PodSpec{Containers: []corev1.Container{{
-				Name:  fmt.Sprintf("container-%s", name),
-				Image: "ignore",
-			}},
-				NodeName: hostname,
-			},
-			Status: corev1.PodStatus{
-				Phase: corev1.PodRunning,
-				Conditions: []corev1.PodCondition{{
-					Type:   corev1.PodScheduled,
-					Status: corev1.ConditionTrue,
-				}},
-				PodIP:  ip.String(),
-				PodIPs: []corev1.PodIP{{IP: ip.String()}},
-			},
-		}
-		return p
-	}
-
-	sendNUpdates := func(n int) map[string]api.Update {
-		expectedEndState := map[string]api.Update{}
-		h.Decoupler.OnStatusUpdated(api.ResyncInProgress)
-		conv := conversion.NewConverter()
-		for i := 0; i < n; i++ {
-			pod := generatePod(n)
-			weps, err := conv.PodToWorkloadEndpoints(pod)
-			Expect(err).NotTo(HaveOccurred())
-			update := api.Update{
-				KVPair:     *weps[0],
-				UpdateType: api.UpdateTypeKVNew,
-			}
-			path, err := model.KeyToDefaultPath(update.Key)
-			Expect(err).NotTo(HaveOccurred())
-			expectedEndState[path] = update
-			h.Decoupler.OnUpdates([]api.Update{update})
-		}
-		return expectedEndState
-	}
-
-	sendNUpdatesThenInSync := func(n int) map[string]api.Update {
-		expectedEndState := sendNUpdates(n)
-		h.Decoupler.OnStatusUpdated(api.InSync)
-		return expectedEndState
-	}
-
 	Describe("with a client connection", func() {
 		var clientCancel context.CancelFunc
 		var recorder *StateRecorder
@@ -296,7 +190,7 @@ var _ = Describe("With an in-process Server", func() {
 
 		// expectFelixClientState asserts that the client eventually reaches the given state.  Then, it
 		// simulates a second connection and check that that also converges to the given state.
-		expectClientState := func(c *clientState, status api.SyncStatus, kvs map[string]api.Update) {
+		expectClientState := func(c *ClientState, status api.SyncStatus, kvs map[string]api.Update) {
 			// Wait until we reach that state.
 			Eventually(c.recorder.Status).Should(Equal(status))
 			Eventually(c.recorder.KVs).Should(Equal(kvs))
@@ -389,7 +283,7 @@ var _ = Describe("With an in-process Server", func() {
 		})
 
 		Describe("with a BGP client", func() {
-			var bgpClient *clientState
+			var bgpClient *ClientState
 
 			BeforeEach(func() {
 				bgpClient = h.CreateClient("bgp", syncproto.SyncerTypeBGP)
@@ -495,7 +389,7 @@ var _ = Describe("With an in-process Server", func() {
 		})
 
 		It("should pass through many KVs", func() {
-			expectedEndState := sendNUpdatesThenInSync(1000)
+			expectedEndState := h.SendInitialSnapshotPods(1000)
 			expectFelixClientState(api.InSync, expectedEndState)
 		})
 
@@ -513,29 +407,16 @@ var _ = Describe("With an in-process Server", func() {
 		var expectedEndState map[string]api.Update
 		BeforeEach(func() {
 			log.SetLevel(log.InfoLevel)
-			expectedEndState = sendNUpdatesThenInSync(500000)
-			h.CreateClients(1)
+			expectedEndState = h.SendInitialSnapshotPods(500000)
+			// The snapshot is huge, so we only create one real client (which records the
+			// keys/values that it sees) and a bunch of no-op clients, which run the protocol
+			// but don't record anything.
 			h.CreateNoOpClients(10)
+			h.CreateClients(1)
 		})
 
-		// expectClientState asserts that every client eventually reaches the given state.
-		expectClientStates := func(status api.SyncStatus, kvs map[string]api.Update) {
-			var wg sync.WaitGroup
-			for _, s := range h.ClientStates {
-				wg.Add(1)
-				go func(s *clientState) {
-					defer wg.Done()
-					defer GinkgoRecover()
-					// Wait until we reach that state.
-					Eventually(s.recorder.Status, 60*time.Second, 50*time.Millisecond).Should(Equal(status))
-					Eventually(s.recorder.KVCompareFn(kvs), 20*time.Second, 200*time.Millisecond).ShouldNot(HaveOccurred())
-				}(s)
-			}
-			wg.Wait()
-		}
-
 		It("should pass through many KVs", func() {
-			expectClientStates(api.InSync, expectedEndState)
+			h.ExpectAllClientsToReachState(api.InSync, expectedEndState)
 		})
 	})
 
@@ -545,20 +426,11 @@ var _ = Describe("With an in-process Server", func() {
 			h.CreateClients(100)
 		})
 
-		// expectClientState asserts that every client eventually reaches the given state.
-		expectClientStates := func(status api.SyncStatus, kvs map[string]api.Update) {
-			for _, s := range h.ClientStates {
-				// Wait until we reach that state.
-				Eventually(s.recorder.Status, 10*time.Second, 200*time.Millisecond).Should(Equal(status))
-				Eventually(s.recorder.KVs, 10*time.Second).Should(Equal(kvs))
-			}
-		}
-
 		It("should drop expected number of connections", func() {
 			// Start a goroutine to watch each client and send us a message on the channel when it stops.
 			finishedC := make(chan int)
 			for _, s := range h.ClientStates {
-				go func(s *clientState) {
+				go func(s *ClientState) {
 					s.client.Finished.Wait()
 					finishedC <- 1
 				}(s)
@@ -596,7 +468,7 @@ var _ = Describe("With an in-process Server", func() {
 			h.Decoupler.OnStatusUpdated(api.ResyncInProgress)
 			h.Decoupler.OnUpdates([]api.Update{configFoobarBazzBiff})
 			h.Decoupler.OnStatusUpdated(api.InSync)
-			expectClientStates(
+			h.ExpectAllClientsToReachState(
 				api.InSync,
 				map[string]api.Update{
 					"/calico/v1/config/foobar": configFoobarBazzBiff,
@@ -605,8 +477,8 @@ var _ = Describe("With an in-process Server", func() {
 		})
 
 		It("should pass through many KVs", func() {
-			expectedEndState := sendNUpdatesThenInSync(1000)
-			expectClientStates(api.InSync, expectedEndState)
+			expectedEndState := h.SendInitialSnapshotPods(1000)
+			h.ExpectAllClientsToReachState(api.InSync, expectedEndState)
 		})
 
 		It("should report the correct number of connections", func() {
@@ -627,7 +499,7 @@ var _ = Describe("With an in-process Server", func() {
 			var wg sync.WaitGroup
 			wg.Add(1)
 			go func() {
-				sendNUpdatesThenInSync(1000)
+				h.SendInitialSnapshotPods(1000)
 				wg.Done()
 			}()
 			defer wg.Wait()
@@ -643,7 +515,7 @@ var _ = Describe("With an in-process Server", func() {
 var _ = Describe("with 5 client connections", func() {
 	const numClients = 5
 
-	var h *serverHarness
+	var h *ServerHarness
 
 	BeforeEach(func() {
 		log.SetLevel(log.InfoLevel) // Debug too verbose for tests with a few clients.
@@ -717,7 +589,7 @@ var _ = Describe("With an in-process Server with short ping timeout", func() {
 	//                      decoupler        filter        cache
 	//
 	var (
-		h *serverHarness
+		h *ServerHarness
 	)
 
 	BeforeEach(func() {
@@ -958,7 +830,7 @@ var _ = Describe("With an in-process Server with long ping interval", func() {
 	//                      decoupler        filter        cache
 	//
 	var (
-		h *serverHarness
+		h *ServerHarness
 	)
 
 	BeforeEach(func() {
@@ -966,6 +838,7 @@ var _ = Describe("With an in-process Server with long ping interval", func() {
 		log.SetLevel(log.DebugLevel)
 		h = NewHarness()
 		h.Config = syncserver.Config{
+			// Effectively disable pings so that we can test the non-ping timeouts.
 			PingInterval: 10000 * time.Second,
 			PongTimeout:  50000 * time.Second,
 			Port:         syncserver.PortRandom,
@@ -1019,110 +892,62 @@ var _ = Describe("With an in-process Server with long ping interval", func() {
 
 var _ = Describe("With an in-process Server with short grace period", func() {
 	var (
-		cacheCxt     context.Context
-		cacheCancel  context.CancelFunc
-		cache        *snapcache.Cache
-		server       *syncserver.Server
-		serverCxt    context.Context
-		serverCancel context.CancelFunc
-		serverAddr   string
-		updIdx       int
+		h *ServerHarness
 	)
 
 	BeforeEach(func() {
-		// Debug logs slow these tests down too much.
-		log.SetLevel(log.InfoLevel)
+		// Default to debug but some more aggressive tests override this below.
+		log.SetLevel(log.DebugLevel)
+		h = NewHarness()
 
-		cache = snapcache.New(snapcache.Config{
-			// Set the batch size small so we can force new Breadcrumbs easily.
-			MaxBatchSize: 10,
-			// Reduce the wake up interval from the default to give us faster tear down.
-			WakeUpInterval: 50 * time.Millisecond,
-		})
-		server = syncserver.New(
-			map[syncproto.SyncerType]syncserver.BreadcrumbProvider{syncproto.SyncerTypeFelix: cache},
-			syncserver.Config{
-				PingInterval:                   10000 * time.Second,
-				PongTimeout:                    50000 * time.Second,
-				Port:                           syncserver.PortRandom,
-				DropInterval:                   1 * time.Second,
-				MaxFallBehind:                  1 * time.Second,
-				NewClientFallBehindGracePeriod: 1 * time.Second,
-				// The snapshot is >10MB; set the write buffer to something much less than that since these
-				// tests need to cause backpressure on Typha.
-				WriteBufferSize: 1024 * 256,
-				DebugLogWrites:  true,
-			})
-		cacheCxt, cacheCancel = context.WithCancel(context.Background())
-		cache.Start(cacheCxt)
-		serverCxt, serverCancel = context.WithCancel(context.Background())
-		server.Start(serverCxt)
-		serverAddr = fmt.Sprintf("127.0.0.1:%d", server.Port())
-		updIdx = 0
+		// Effectively disable pings, so we can hit the grace period timers.
+		h.Config.PingInterval = 10000 * time.Second
+		h.Config.PongTimeout = 50000 * time.Second
+
+		// Short timeouts so we can hit them quickly.
+		h.Config.MaxFallBehind = time.Second
+		h.Config.NewClientFallBehindGracePeriod = time.Second
+
+		// The snapshot is >10MB; set the write buffer to something much less than that since these
+		// tests need to cause backpressure on Typha.
+		h.Config.WriteBufferSize = 1024 * 256
+		// Since we rely on back-pressure, it's really handy to have a log of exactly what writes
+		// happen and when.
+		h.Config.DebugLogWrites = true
+
+		h.Start()
 	})
 
 	AfterEach(func() {
-		if server != nil {
-			serverCancel()
-			log.Info("Waiting for server to shut down")
-			server.Finished.Wait()
-			log.Info("Done waiting for server to shut down")
-		}
-		if cache != nil {
-			cacheCancel()
-		}
+		h.Stop()
 	})
-
-	sendNUpdates := func(n int) map[string]api.Update {
-		expectedEndState := map[string]api.Update{}
-		cache.OnStatusUpdated(api.ResyncInProgress)
-		for i := 0; i < n; i++ {
-			update := api.Update{
-				KVPair: model.KVPair{
-					Key: model.GlobalConfigKey{
-						Name: fmt.Sprintf("foo%v", i+updIdx),
-					},
-					// Nice big value so that we can fill up the send queue.
-					Value:    randomValue(i),
-					Revision: fmt.Sprintf("%v", i+updIdx),
-				},
-				UpdateType: api.UpdateTypeKVNew,
-			}
-			path, err := model.KeyToDefaultPath(update.Key)
-			Expect(err).NotTo(HaveOccurred())
-			expectedEndState[path] = update
-			cache.OnUpdates([]api.Update{update})
-			updIdx += n
-		}
-		return expectedEndState
-	}
-
-	sendNUpdatesThenInSync := func(n int) map[string]api.Update {
-		expectedEndState := sendNUpdates(n)
-		cache.OnStatusUpdated(api.InSync)
-		return expectedEndState
-	}
 
 	Describe("with lots of KVs", func() {
 		const initialSnapshotSize = 10000
-		BeforeEach(func() {
-			sendNUpdatesThenInSync(initialSnapshotSize)
-		})
 
-		AfterEach(func() {
+		logStats := func(note string) {
 			for _, stat := range []string{
 				"typha_snapshot_generated",
 			} {
 				value, _ := getPerSyncerCounter(syncproto.SyncerTypeFelix, stat)
-				log.Infof("Counter: %s =  %v", stat, int(value))
+				log.Infof("%s: counter: %s =  %v", note, stat, int(value))
 			}
 			for _, stat := range []string{
 				"typha_snapshot_raw_bytes",
 				"typha_snapshot_compressed_bytes",
 			} {
 				value, _ := getPerSyncerGauge(syncproto.SyncerTypeFelix, stat)
-				log.Infof("Gauge: %s =  %v", stat, int(value))
+				log.Infof("%s: gauge: %s =  %v", note, stat, int(value))
 			}
+		}
+
+		BeforeEach(func() {
+			logStats("Start of test")
+			h.SendInitialSnapshotConfigs(initialSnapshotSize)
+		})
+
+		AfterEach(func() {
+			logStats("End of test")
 		})
 
 		It("client should get a grace period after reading the snapshot", func() {
@@ -1139,7 +964,7 @@ var _ = Describe("With an in-process Server with short grace period", func() {
 			recorder.BlockAfterNUpdates(1, 2500*time.Millisecond)
 
 			client := syncclient.New(
-				[]discovery.Typha{{Addr: serverAddr}},
+				[]discovery.Typha{{Addr: h.Addr()}},
 				"test-version",
 				"test-host",
 				"test-info",
@@ -1171,13 +996,13 @@ var _ = Describe("With an in-process Server with short grace period", func() {
 			log.Info("Sleeping 1s")
 			time.Sleep(time.Second)
 			log.Info("Sending 1 update")
-			sendNUpdates(1)
+			h.SendConfigUpdates(1)
 			log.Info("Sent 1 update")
 			// Make a breadcrumb 1s later.
 			log.Info("Sleeping 1s")
 			time.Sleep(time.Second)
 			log.Info("Sending 1 update")
-			sendNUpdates(1)
+			h.SendConfigUpdates(1)
 			log.Info("Sent 1 update")
 
 			// Client should wake up around now and start catching up.
@@ -1186,7 +1011,7 @@ var _ = Describe("With an in-process Server with short grace period", func() {
 			log.Info("Sleeping 1s")
 			time.Sleep(time.Second)
 			log.Info("Sending 1 update")
-			sendNUpdates(1)
+			h.SendConfigUpdates(1)
 			log.Info("Sent 1 update")
 
 			Eventually(recorder.Len, 2*time.Second).Should(BeNumerically("==", initialSnapshotSize+3))
@@ -1206,7 +1031,7 @@ var _ = Describe("With an in-process Server with short grace period", func() {
 			recorder.BlockAfterNUpdates(initialSnapshotSize+1, 2500*time.Millisecond)
 
 			client := syncclient.New(
-				[]discovery.Typha{{Addr: serverAddr}},
+				[]discovery.Typha{{Addr: h.Addr()}},
 				"test-version",
 				"test-host",
 				"test-info",
@@ -1227,17 +1052,17 @@ var _ = Describe("With an in-process Server with short grace period", func() {
 			log.SetLevel(log.DebugLevel)
 			// Make a breadcrumb.
 			time.Sleep(time.Second)
-			sendNUpdates(initialSnapshotSize)
+			h.SendConfigUpdates(initialSnapshotSize)
 
 			// Client should read the first update from the above and then block.
 
 			// Make a breadcrumb 1s later.
 			time.Sleep(time.Second)
-			sendNUpdates(1)
+			h.SendConfigUpdates(1)
 
 			// Make a breadcrumb 1s later.
 			time.Sleep(time.Second)
-			sendNUpdates(1)
+			h.SendConfigUpdates(1)
 
 			// Client should wake up around now but be too far behind and get disconnected.
 
@@ -1258,17 +1083,6 @@ var _ = Describe("With an in-process Server with short grace period", func() {
 		})
 	})
 })
-
-func randomValue(n int) string {
-	v := fmt.Sprint(n, "=")
-	buf := make([]byte, 1000)
-	rand.Read(buf)
-	var outBuf bytes.Buffer
-	enc := base64.NewEncoder(base64.StdEncoding, &outBuf)
-	_, _ = enc.Write(buf)
-	v += outBuf.String()
-	return v
-}
 
 func getGlobalGauge(name string) (float64, error) {
 	mfs, err := prometheus.DefaultGatherer.Gather()
@@ -1678,154 +1492,3 @@ var _ = Describe("with server requiring TLS", func() {
 		It("TLS connection with good CN and URI should fail", testTLSGoodCNURI(false))
 	})
 })
-
-// serverHarness runs a syncserver.Server with a couple of cache syncer cache types and allows for creating test
-// clients.  The server configuration can be adjusted after creation but before calling Start(). Stop() should be
-// called in an AfterEach.  The server runs on a random high numbered port.
-type serverHarness struct {
-	Decoupler, BGPDecoupler *calc.SyncerCallbacksDecoupler
-	ValFilter               *calc.ValidationFilter
-	cacheCxt                context.Context
-	cacheCancel             context.CancelFunc
-	FelixCache, BGPCache    *snapcache.Cache
-	Server                  *syncserver.Server
-	serverCxt               context.Context
-	ServerCancel            context.CancelFunc
-	Config                  syncserver.Config
-
-	ClientStates     []*clientState
-	NoOpClientStates []*clientState
-}
-
-type clientState struct {
-	clientCxt    context.Context
-	clientCancel context.CancelFunc
-	client       *syncclient.SyncerClient
-	recorder     *StateRecorder
-	syncerType   syncproto.SyncerType
-}
-
-func (h *serverHarness) Start() {
-	h.Server = syncserver.New(
-		map[syncproto.SyncerType]syncserver.BreadcrumbProvider{
-			syncproto.SyncerTypeFelix: h.FelixCache,
-			syncproto.SyncerTypeBGP:   h.BGPCache,
-		},
-		h.Config)
-
-	go h.Decoupler.SendToContext(h.cacheCxt, h.ValFilter)
-	go h.BGPDecoupler.SendToContext(h.cacheCxt, h.BGPCache)
-	h.FelixCache.Start(h.cacheCxt)
-	h.BGPCache.Start(h.cacheCxt)
-	h.serverCxt, h.ServerCancel = context.WithCancel(context.Background())
-	h.Server.Start(h.serverCxt)
-}
-
-func (h *serverHarness) Addr() string {
-	return fmt.Sprintf("127.0.0.1:%d", h.Server.Port())
-}
-
-func (h *serverHarness) Stop() {
-	allClients := append(h.ClientStates, h.NoOpClientStates...)
-
-	for _, c := range allClients {
-		c.clientCancel()
-	}
-	for _, c := range allClients {
-		if c.client != nil {
-			log.Info("Waiting for client to shut down.")
-			c.client.Finished.Wait()
-			log.Info("Done waiting for client to shut down.")
-		}
-	}
-
-	h.ServerCancel()
-	log.Info("Waiting for server to shut down")
-	h.Server.Finished.Wait()
-	log.Info("Done waiting for server to shut down")
-	h.cacheCancel()
-}
-
-func (h *serverHarness) CreateNoOpClient(id interface{}, syncType syncproto.SyncerType) *clientState {
-	c := h.createClient(id, syncType, NoOpCallbacks{})
-	h.NoOpClientStates = append(h.ClientStates, c)
-	return c
-}
-
-func (h *serverHarness) CreateClient(id interface{}, syncType syncproto.SyncerType) *clientState {
-	recorder := NewRecorder()
-	c := h.createClient(id, syncType, recorder)
-	c.recorder = recorder
-	go recorder.Loop(c.clientCxt)
-	h.ClientStates = append(h.ClientStates, c)
-	return c
-}
-
-func (h *serverHarness) createClient(id interface{}, syncType syncproto.SyncerType, callbacks api.SyncerCallbacks) *clientState {
-	clientCxt, clientCancel := context.WithCancel(context.Background())
-	serverAddr := fmt.Sprintf("127.0.0.1:%d", h.Server.Port())
-	client := syncclient.New(
-		[]discovery.Typha{{Addr: serverAddr}},
-		"test-version",
-		fmt.Sprintf("test-host-%v", id),
-		"test-info",
-		callbacks,
-		&syncclient.Options{
-			SyncerType: syncType,
-		},
-	)
-
-	err := client.Start(clientCxt)
-	Expect(err).NotTo(HaveOccurred())
-
-	cs := &clientState{
-		clientCxt:    clientCxt,
-		client:       client,
-		clientCancel: clientCancel,
-		syncerType:   syncType,
-	}
-	return cs
-}
-
-func (h *serverHarness) CreateNoOpClients(n int) {
-	for i := 0; i < n; i++ {
-		h.CreateNoOpClient(i, syncproto.SyncerTypeFelix)
-	}
-}
-
-func (h *serverHarness) CreateClients(n int) {
-	for i := 0; i < n; i++ {
-		h.CreateClient(i, syncproto.SyncerTypeFelix)
-	}
-}
-
-func NewHarness() *serverHarness {
-	// Set up a pipeline:
-	//
-	//    This goroutine -> callback -chan-> validation -> snapshot -> server
-	//                      decoupler        filter        cache
-	//
-	h := &serverHarness{}
-	h.Decoupler = calc.NewSyncerCallbacksDecoupler()
-	h.FelixCache = snapcache.New(snapcache.Config{
-		// Set the batch size small so we can force new Breadcrumbs easily.
-		MaxBatchSize: 10,
-		// Reduce the wake up interval from the default to give us faster tear down.
-		WakeUpInterval: 50 * time.Millisecond,
-	})
-	h.BGPDecoupler = calc.NewSyncerCallbacksDecoupler()
-	h.BGPCache = snapcache.New(snapcache.Config{
-		// Set the batch size small so we can force new Breadcrumbs easily.
-		MaxBatchSize: 10,
-		// Reduce the wake up interval from the default to give us faster tear down.
-		WakeUpInterval: 50 * time.Millisecond,
-	})
-	h.cacheCxt, h.cacheCancel = context.WithCancel(context.Background())
-	h.ValFilter = calc.NewValidationFilter(h.FelixCache)
-	h.Config = syncserver.Config{
-		PingInterval: 10 * time.Second,
-		Port:         syncserver.PortRandom,
-		DropInterval: 50 * time.Millisecond,
-	}
-	return h
-}
